Integrate the index into the main code base
