+++ Makefile
@@ -5,6 +5,9 @@ ALEMBIC=alembic -c opensanctions/migrate/alembic.ini
 
 all: run
 
+db:
+	curl -o data/state/opensanctions.sqlite https://data.opensanctions.org/state/opensanctions.sqlite
+
 workdir:
 	mkdir -p data/postgres
